Code cleanup
Took 5 minutes
