+++ common/src/main/java/me/tuhin47/exporter/ExcelGenerator.java
@@ -1,4 +1,4 @@
-package me.tuhin47.utils;
+package me.tuhin47.exporter;
 
 import lombok.extern.log4j.Log4j2;
 import org.apache.poi.ss.usermodel.Cell;
