+++ common/src/main/java/me/tuhin47/exporter/ExportColumn.java
@@ -1,4 +1,5 @@
-package me.tuhin47.utils;
+package me.tuhin47.exporter;
+
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,7 +7,13 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface ExcelSheet {
-    String value() default "sheet 1";
+@Target(ElementType.FIELD)
+public @interface ExportColumn {
+
+    String value();
+
+    /**
+     * @return 0 based sort order
+     */
+    int sortOrder() default -1;
 }
