+++ productservice/src/main/java/com/microservice/productservice/conttoller/Impl/ProductControllerImpl.java
@@ -7,7 +7,7 @@ import com.microservice.productservice.service.ProductService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import me.tuhin47.searchspec.SearchCriteria;
-import me.tuhin47.utils.ExcelGenerator;
+import me.tuhin47.exporter.ExcelGenerator;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
