+++ productservice/src/main/java/com/microservice/productservice/payload/response/ProductResponse.java
@@ -4,9 +4,9 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import me.tuhin47.utils.ExcelDTO;
-import me.tuhin47.utils.ExcelSheet;
-import me.tuhin47.utils.ExportColumn;
+import me.tuhin47.exporter.ExcelDTO;
+import me.tuhin47.exporter.ExcelSheet;
+import me.tuhin47.exporter.ExportColumn;
 
 @Data
 @Builder
